move transform function into AddMergeTemplates
